security/requirements_sec.txt
semantic-kernel==1.18.2
asyncio==3.4.3
python-dotenv==1.0.1
llm-guard==0.3.15